backend/requirements.txt
django==3.2
psycopg2-binary==2.8.6
Pillow==8.2.0
djangorestframework==3.12.4
django-phonenumber-field[phonenumberslite]==5.0.0
django-cors-headers==3.7.0
pytest==6.2.3
pytest-django==4.1.0
python-dateutil==2.8.1